+++ src/main/java/com/minswap/hrms/repsotories/RequestTypeRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface RequestTypeRepository extends JpaRepository<RequestType,Integer> {
+public interface RequestTypeRepository extends JpaRepository<RequestType,Long> {
 
     @Query("select rt.requestTypeId from RequestType rt")
     List<Long> getAllRequestTypeId();
